# web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheoson API",
    "version": "2.11.0"
  },
  "paths": {
...
    "/api/downloads": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "List Downloads",
        "operationId": "list_downloads_api_downloads_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DownloadJobSchema"
                  },
                  "type": "array",
                  "title": "Response List Downloads Api Downloads Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Start Download",
        "operationId": "start_download_api_downloads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/downloads/{job_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Download",
        "operationId": "get_download_api_downloads__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "downloads"
        ],
        "summary": "Delete Download",
        "operationId": "delete_download_api_downloads__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/downloads/{job_id}/cancel": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Cancel Download",
        "operationId": "cancel_download_api_downloads__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/downloads/{job_id}/retry": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Retry Download",
        "operationId": "retry_download_api_downloads__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadJobSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/downloads/batch": {
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Batch Download",
        "description": "Start multiple downloads at once (max 20).",
        "operationId": "batch_download_api_downloads_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DownloadJobSchema"
                  },
                  "type": "array",
                  "title": "Response Batch Download Api Downloads Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchDownloadRequest": {
        "properties": {
          "track_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Track Ids"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "mp3"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "320"
          },
          "embed_artwork": {
            "type": "boolean",
            "title": "Embed Artwork",
            "default": true
          },
          "embed_lyrics": {
            "type": "boolean",
            "title": "Embed Lyrics",
            "default": true
          },
          "embed_metadata": {
            "type": "boolean",
            "title": "Embed Metadata",
            "default": true
          },
          "file_naming": {
            "type": "string",
            "title": "File Naming",
            "default": "artist-title"
          },
          "custom_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Path"
          },
          "retries": {
            "type": "integer",
            "title": "Retries",
            "default": 3
          },
          "speed_limit": {
            "type": "integer",
            "title": "Speed Limit",
            "default": 0
          },
          "concurrency": {
            "type": "integer",
            "title": "Concurrency",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "track_ids"
        ],
        "title": "BatchDownloadRequest"
      },
...
      "DownloadJobSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "trackId": {
            "type": "string",
            "title": "Trackid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "artist": {
            "type": "string",
            "title": "Artist",
            "default": ""
          },
          "artworkUrl": {
            "type": "string",
            "title": "Artworkurl",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "searching",
              "downloading",
              "converting",
              "tagging",
              "done",
              "error",
              "cancelled"
            ],
            "title": "Status",
            "default": "queued"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "format": {
            "type": "string",
            "enum": [
              "mp3",
              "flac",
              "opus",
              "m4a",
              "wav"
            ],
            "title": "Format",
            "default": "mp3"
          },
          "quality": {
            "type": "string",
            "enum": [
              "128",
              "192",
              "256",
              "320",
              "best"
            ],
            "title": "Quality",
            "default": "320"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filepath"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "error",
          "filePath",
          "id"
        ],
        "title": "DownloadJobSchema"
      },
      "DownloadRequestSchema": {
        "properties": {
          "trackId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackid"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "format": {
            "type": "string",
            "enum": [
              "mp3",
              "flac",
              "opus",
              "m4a",
              "wav"
            ],
            "title": "Format",
            "default": "mp3"
          },
          "quality": {
            "type": "string",
            "enum": [
              "128",
              "192",
              "256",
              "320",
              "best"
            ],
            "title": "Quality",
            "default": "320"
          },
          "embedArtwork": {
            "type": "boolean",
            "title": "Embedartwork",
            "default": true
          },
          "embedLyrics": {
            "type": "boolean",
            "title": "Embedlyrics",
            "default": true
          },
          "embedMetadata": {
            "type": "boolean",
            "title": "Embedmetadata",
            "default": true
          },
          "retries": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Retries",
            "default": 3
          },
          "speedLimit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Speedlimit",
            "default": 0
          },
          "concurrency": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Concurrency",
            "default": 3
          },
          "fileNaming": {
            "type": "string",
            "enum": [
              "artist-title",
              "title-artist",
              "id"
            ],
            "title": "Filenaming",
            "default": "artist-title"
          },
          "customPath": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Custompath"
          }
        },
        "type": "object",
        "title": "DownloadRequestSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}